Refactor metrics backend configuration and update documentation
- Changed default metrics backend from Prometheus to VictoriaMetrics in deploy-to-kind.sh.
- Updated Helm repository addition logic for VictoriaMetrics.
- Modified installation process for VictoriaMetrics stack and vmagent to handle failures gracefully.
- Removed legacy vmagent-values.yaml and integrated its configuration into the main deployment.
- Added new ENVIRONMENT_VARIABLES.md to document environment variable configurations.
- Updated backend-deployment.yaml to include new metrics backend environment variables.
- Changed Grafana admin password from 'pod-metrics-admin' to 'admin' for consistency.
- Removed unused victoriametrics-values.yaml file.
- Updated vmagent-values.yaml to streamline configuration and improve scraping for kube-state-metrics.
